Fix: Creation of form validation

+++ MyTraining/Workout/View/WorkoutCardView.swift
@@ -72,7 +72,9 @@ struct WorkoutCardView: View {
                             
                             Button{
                                 workoutViewModel.deleteWorkout(workoutId: id)
-                                chartViewModel.deleteChart(chartId: chartViewModel.getChartByWorkoutId(workoutId: id).id)
+                                if chartViewModel.isWorkoutExists(workoutId: id){
+                                    chartViewModel.deleteChart(chartId: chartViewModel.getChartByWorkoutId(workoutId: id).id)
+                                }
                             } label: {
                                 Image(systemName: "trash")
                                     .foregroundColor(.red)
